extracting stats page
