+++ app/Http/Controllers/V2ExperimentsController.php
@@ -2,48 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use DB;
-use Carbon\Carbon;
-use Illuminate\Support\Collection;
-use App\User;
 use App\Content;
 use App\Destination;
 
 class V2ExperimentsController extends Controller
 {
-    public function getMonthlyStat($model)
-    {
-        return Collection::times(3, function($year) use ($model) {
-            $model = 'App\\'.$model;
-            $table = (new $model)->getTable();
-            return $model::select(
-                    DB::raw("DATE_FORMAT(created_at, '%M') date"),
-                    DB::raw('count('.$table.'.id) as aggregate')
-                )
-                ->groupBy(DB::raw('MONTH('.$table.'.created_at)'))
-                ->whereBetween(
-                    'created_at',
-                    $year - 1 == 0 ? [
-                        Carbon::now()->startOfYear(),
-                        Carbon::now()->startOfMonth()
-                    ] : [
-                        Carbon::now()->subYears($year - 1)->startOfYear(),
-                        Carbon::now()->subYears($year - 1)->endOfYear()
-                    ]
-                )
-                ->orderBy('created_at')
-                ->pluck('aggregate');
-        })
-        ->map(function($values, $year) {
-            return collect()
-                // Replace with ->pad in Laravel 5.5
-                ->put('values', array_pad($values->all(), 12, 0))
-                ->put('title', Carbon::now()
-                    ->subYears($year)
-                    ->format('Y')
-                ); 
-        });
-    }
 
     public function index()
     {
@@ -52,23 +15,6 @@ class V2ExperimentsController extends Controller
         return layout('1col')
 
             ->with('content', collect()
-
-                ->push(component('Title')
-                    ->with('title', 'Linecart')
-                )
-
-                ->merge(collect(['User', 'Content', 'Comment', 'Flag'])
-                    ->flatMap(function($model) {
-                        return collect()
-                            ->push(component('Title')
-                                ->is('small')
-                                ->with('title', $model)
-                            )
-                            ->push(component('Linechart')
-                                ->with('items', $this->getMonthlyStat($model))
-                            );
-                    })
-                )
  
                 ->push(component('Title')
                     ->with('title', 'Barchart')
